Appeased linter on LandingPage.jsx
